N2 has wonky partial encryption when writing
There is a partial encryption on data written to the N2, which results in the properties view being corrupted when not decrypting data or a decryption failure. The best compromise is to check the encryption and display the data for the user to adjust.

Verify ALL data is stripped, since encryption breaks N2