Support closing N2 viewer with back button
Restore the scaling factor with fixed refreshing

Note: If you need scaling, try the default. If the default doesn't work, try a lower number for less zoom and a smaller number for more zoom. You must exit settings to see the update (requirement to maintain stability). If scaling is not something you prefer, do NOT touch this toggle.

Visual notice of double scan, Settings cleanup