Begin implementing Storage Access Framework
Divert obscure storage to Android handlers

Import some fallback methods for Storage access

As documented in Chainfire's misguided attempt to spring the issue on Google during a conference (https://www.xda-developers.com/android-3-2-code-inadvertently-preventing-write-access-to-external-storage/), media storage was fundamentally inaccessible to third party apps because of a permission typo. Samsung devices attempted to work around this with a similar redirect, which likely inspired the original code that was rewritten and submitted (by myself) for use in reicast and other emulator projects to gather external files.