Set view visibility on the UI thread (Fixes #379)
Discard a relatively pointless automation (Fixes #380)

Fix the UI, despite the implication of it appearing

This UI should really never be reached, unless setup has not been completed properly and the keys don't exist. It also shouldn't crash to avoid disguising the real issue of no keys.

Urge issues to be posted on GitHub, not Amiibomb

Since the project is not hosted on Amiibomb, the moderators of Amiibomb are unresponsive, and the community usually has counterproductive advice, let's urge posting issues where they have a legitimate chance of being addressed.

Fix an issue with the list not appearing on first load

Restore view update order (Fixes #381)