Move folder selection to button, default settings
Folder settings will always be displayed when clicking the folder icon. Settings will always be displayed when clicking the settings menu item. When the drawer is closed and slid open, it will always display settings. This resolves a somewhat confusing design meant to make both accessible in the limited action space provided.